Fix Device Stats object constructor when there is no event

+++ watergate_local_api/models/auto_shut_off_report.py
@@ -8,7 +8,7 @@ class AutoShutOffReport:
 
     def __init__(
         self,
-        timestamp: float,
+        timestamp: int,
         type: str,
         duration: int,
         volume: int,
